# Implement attention cache in Transformer for faster decoding