--- a/espnet/nets/pytorch_backend/transformer/decoder_layer.py
+++ b/espnet/nets/pytorch_backend/transformer/decoder_layer.py
@@ -7,7 +7,6 @@
 """Decoder self-attention layer definition."""
 
 import torch
-
 from torch import nn
 
 from espnet.nets.pytorch_backend.transformer.layer_norm import LayerNorm
@@ -47,22 +46,36 @@ class DecoderLayer(nn.Module):
             self.concat_linear1 = nn.Linear(size + size, size)
             self.concat_linear2 = nn.Linear(size + size, size)
 
-    def forward(self, tgt, tgt_mask, memory, memory_mask):
+    def forward(self, tgt, tgt_mask, memory, memory_mask, cache=None):
         """Compute decoded features.
 
-        :param torch.Tensor tgt: decoded previous target features (batch, max_time_out, size)
-        :param torch.Tensor tgt_mask: mask for x (batch, max_time_out)
-        :param torch.Tensor memory: encoded source features (batch, max_time_in, size)
-        :param torch.Tensor memory_mask: mask for memory (batch, max_time_in)
+        Args:
+            tgt (torch.Tensor): decoded previous target features (batch, max_time_out, size)
+            tgt_mask (torch.Tensor): mask for x (batch, max_time_out)
+            memory (torch.Tensor): encoded source features (batch, max_time_in, size)
+            memory_mask (torch.Tensor): mask for memory (batch, max_time_in)
+            cache (torch.Tensor): cached output (batch, max_time_out-1, size)
+
         """
-        residual = tgt
+        if cache is None:
+            tgt_q = tgt
+            tgt_q_mask = tgt_mask
+        else:
+            # compute only the last frame query keeping dim: max_time_out -> 1
+            assert cache.shape == (tgt.shape[0], tgt.shape[1] - 1, self.size)
+            tgt_q = tgt[:, -1:, :]
+            tgt_q_mask = None
+            if tgt_mask is not None:
+                tgt_q_mask = tgt_mask[:, -1:, :]
+
+        residual = tgt_q
         if self.normalize_before:
-            tgt = self.norm1(tgt)
+            tgt_q = self.norm1(tgt_q)
         if self.concat_after:
-            tgt_concat = torch.cat((tgt, self.self_attn(tgt, tgt, tgt, tgt_mask)), dim=-1)
+            tgt_concat = torch.cat((tgt_q, self.self_attn(tgt_q, tgt, tgt, tgt_q_mask)), dim=-1)
             x = residual + self.concat_linear1(tgt_concat)
         else:
-            x = residual + self.dropout(self.self_attn(tgt, tgt, tgt, tgt_mask))
+            x = residual + self.dropout(self.self_attn(tgt_q, tgt, tgt, tgt_q_mask))
         if not self.normalize_before:
             x = self.norm1(x)
 
@@ -83,5 +96,8 @@ class DecoderLayer(nn.Module):
         x = residual + self.dropout(self.feed_forward(x))
         if not self.normalize_before:
             x = self.norm3(x)
+
+        if cache is not None:
+            x = torch.cat([cache, x], dim=1)
 
         return x, tgt_mask, memory, memory_mask