- [x] benchmark with synth data
- [x] benchmark with real ASR data
- [x] TTS decoding support