@ShigekiKarita 2019-10-07
![benchmark](https://user-images.githubusercontent.com/6745326/66282653-18155f80-e8fb-11e9-8368-c2c653e95517.png)

I beat synth data. See https://github.com/espnet/espnet/pull/1240/files#diff-8d200e61c2b1cb67a2b4f2d1b408b48c for the script