@ShigekiKarita 2019-10-08
benchmark on librispeech utt (10.44 sec)
```bash
ffmpeg -i ~/work/dataset/LibriSpeech/test-clean/1089/134686/1089-134686-0000.flac  1089-134686-0000.wav

time ../../../utils/recog_wav.sh --models librispeech.transformer.v1 --decode_config $(change_yaml.py conf/decode.yaml -a api=v2 -a lm-weight=0 -a ctc-weight=0 -a beam-size=1) ./1089-134686-0000.wav
```

- baseline: 40.62s user 1.52s system 99% cpu 42.449 total
- this PR: 16.12s user 1.46s system 98% cpu 17.833 total

NOTE: data prep before stage 3 took 5.09s user 1.50s system 96% cpu 6.837 total

Now this PR realize almost realtime in decoding stage (17.833 - 6.837 = 10.99 sec)