@ShigekiKarita 2019-10-08
I will create Transformer TTS support  in another PR because TTS test is broken (fixed in #1244). So I need review now.