@ShigekiKarita 2019-10-08
I added TTS support (it was much easier than I expected)

```bash
echo "THIS IS A DEMONSTRATION OF TEXT TO SPEECH." > example.txt
time ../../../utils/synth_wav.sh --models ljspeech.transformer.v1 example.txt
```

- baseline:  17.58s user 0.17s system 99% cpu 17.870 total
- cached:  9.18s user 0.24s system 98% cpu 9.564 total