@ShigekiKarita 2019-10-08
done but this PR requires #1244 before merging to v.0.6.0 because of CI failures