@ShigekiKarita 2019-10-08
After the review, I found the slight difference in normalization order, and it results differently. Now it is fixed.

ASR result
- v.0.6.0: ▁HE▁HOPED▁THERE▁WOULD▁BE▁STEW▁FOR▁DINNER▁TURNIPS▁AND▁CARROTS▁AND▁BRUISED▁POTATOES▁AND▁FAT▁MUTTON▁PIECES▁TO▁BE▁LADLED▁OUT▁IN▁THICK▁PEPPERED▁FLOUR▁FAT▁AND▁SAUCE
- this PR: ▁HE▁HOPED▁THERE▁WOULD▁BE▁STEW▁FOR▁DINNER▁TURNIPS▁AND▁CARROTS▁AND▁BRUISED▁POTATOES▁AND▁FAT▁MUTTON▁PIECES▁TO▁BE▁LADLED▁OUT▁IN▁THICK▁PEPPERED▁FLOUR▁FAT▁AND▁SAUCE