@mowayao 2019-10-12
I tried https://github.com/ShigekiKarita/espnet/blob/c18e407648e1f706c31ebab9fe012895658df750/test/test_transformer_decode.py, but attention cache in ASR becomes slow when use cuda.
![benchmark_gpu](https://user-images.githubusercontent.com/6467795/66694141-1e348300-ece3-11e9-93c7-28aa6e5cdb05.png)