@mowayao 2019-10-17
> Maybe feat/time is too smaller than cuda overheads, e.g., kernel lauch. CPU/GPU memory transfer. Can you try larger setting? 2019年10月12日(土) 5:27 mowayao <notifications@github.com>:
> […](#)
> I tried https://github.com/ShigekiKarita/espnet/blob/c18e407648e1f706c31ebab9fe012895658df750/test/test_transformer_decode.py, but attention cache in ASR becomes slow when use cuda. [image: benchmark_gpu] <https://user-images.githubusercontent.com/6467795/66694141-1e348300-ece3-11e9-93c7-28aa6e5cdb05.png> — You are receiving this because you authored the thread. Reply to this email directly, view it on GitHub <#1240?email_source=notifications&email_token=ABTOZ3VGG5CP5KZK3LZUAK3QOE7YLA5CNFSM4I564MC2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEBBUJKQ#issuecomment-541279402>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ABTOZ3Q64UPAON5KX4O4YVTQOE7YLANCNFSM4I564MCQ> .

I set adim=256, odim=7000,  linear_units=2048, num_blocks=6,  still got the similar result. 
![benchmark_gpu](https://user-images.githubusercontent.com/6467795/66989906-d1dcaf00-f0f7-11e9-8c67-527c782e98ac.png)